Update CLI. Rename parser
